Committing 'in-flight' work
Quotes b/c I've not really worked on this at all for quite awhile... but trying to pick this back up again.

So I found a decent library for reading & writing WASM files; wasmer wasn't the right fit for that job.  It's more focused on actually executing WASM (fair enough).

So with this commit I can import module files into the sqlite database and "instantiate" objects.

At this point objects are just modules copied byte, for byte.  The only difference (TODO) is that I can send a message to an "object" which will be loaded up into wasmer as an instance.

When that wasm instance finishes executing, I'll write it back to disk but replace the .DATA segment with whatever is in it's memory.

In theory, this will work... but I'm sure there will be edge cases to be handled.  Particularly wrt reference types and component model et al.

But got to start somewhere.