Cleaning up code a bit
Right now we're assuming a specific module has been imported.. for testing purposes.

But if I'm going to be able to import modules generally, it looks like I've got to re-write the module file upon importing.

Among other things...
1. Convert all globals to imported globals
2. Convert all memories to exported memories w/ proper data segments
3. et al

Since I'll need to actually persist the expected values of globals outside of the module, methinks useing the custom segment may be in order.