The beginnings of an ABI
Every instance has an inbox & outbox buffer.  The host sends a message to the instance by resizing the inbox appropriatly with `prepare_inbox`, and writing to the inbox buffer directly.

Then the host invokes `message_received`.  The client gets the correct buffer, based on the information provided earlier via `prepare_inbox`.

The client sends a message to the host by preparing it's outbox buffer.  Since the outbox is a client side buffer, no need to invoke a method in the host.
When the outbox buffer is ready, the client invokes `send_message` with the details of outbox buffers dimensions.