A few more minor fixups
Before I can actually create a native module; I need to refactor
how I've handled the state of the image, namespace, and executing session.

That way, wether a particular object is native or wasm, I just hand it a reference to (image, namespace, session) and it's able to interact with everything.

Right now we're inside out from where we need to be.