More figuring out how 2 async
TLDR
1. Executors just keep a list of things that implement Future (can be polled).
2. When Executors poll a future, they give that future a Waker, which just tells the Executor a given Future is ready to be polled again.
3. You integrate into the host OS's synchronization primitives by creating a type that implements Future.
4. Any function which returns a type that implements Future is async/awaitable (!!)

The top level of an async call stack must be sync code, with a reference to an executor.
How does this tie in with WASM though, given the single threadedness and the fact the executor itself is deferring to the host & we want to resume executing when the host calls back into us?